Added mutation comments.

+++ user_contracts/api/mutations.py
@@ -10,6 +10,13 @@ from .types import UserType, ContractType
 
 
 class CreateUserMutation(graphene.Mutation):
+    """
+    Mutation for creating a new user in the GraphQL API.
+
+    This mutation handles the creation of a new user by accepting an input 
+    of type `UserInput`. It returns the created user object, a success flag, 
+    and a message indicating the result of the operation.
+    """
     class Arguments:
         input = UserInput(required=True)
 
@@ -32,6 +39,13 @@ class CreateUserMutation(graphene.Mutation):
 
 
 class UpdateUserMutation(graphene.Mutation):
+    """
+    Mutation for updating an existing user in the GraphQL API.
+
+    This mutation handles the update of a existing user by accepting an input 
+    of type `UserInput` and `graphene.ID`. It returns the updated user object, a success flag, 
+    and a message indicating the result of the operation.
+    """
     class Arguments:
         id = graphene.ID(required=True)
         input = UserInput(required=True)
@@ -59,6 +73,16 @@ class UpdateUserMutation(graphene.Mutation):
 
 
 class DeleteUserMutation(graphene.Mutation):
+    """
+    Mutation for deleting an existing user in the GraphQL API.
+
+    This mutation handles the delete of a existing user by accepting an input 
+    of type `graphene.ID`. In addiiton, before trying to delete a user, the code
+    will check if the user is attached to a contract. If not, the user will be delete, else
+    a message will be displayed not allowing the user to be removed.
+    It returns the updated user object, a success flag, 
+    and a message indicating the result of the operation.
+    """
     class Arguments:
         id = graphene.ID(required=True)
 
@@ -69,15 +93,31 @@ class DeleteUserMutation(graphene.Mutation):
     def mutate(self, info, id):
         try:
             user = User.objects.get(pk=id)
+            
+            # Check if the user is attached to any contracts
+            if Contract.objects.filter(user=user).exists():
+                raise GraphQLError("Cannot delete user because they are attached to a contract.")
+            
+            # Proceed with deletion if no contracts are found
             user.delete()
             return DeleteUserMutation(
                 success=True, message="User deleted successfully."
             )
+        except User.DoesNotExist:
+            raise GraphQLError(f"User with ID {id} does not exist.")
         except Exception as e:
             raise GraphQLError(f"Could not delete user: {str(e)}")
 
 
 class CreateContractMutation(graphene.Mutation):
+    """
+    Mutation for creating a new contract in the GraphQL API.
+
+    This mutation handles the creation of a new contract by accepting an input 
+    of type `ContractInput` attaching an user to it. 
+    This mutation also returns the created user object, a success flag, 
+    and a message indicating the result of the operation.
+    """
     class Arguments:
         input = ContractInput(required=True)
 
@@ -110,6 +150,13 @@ class CreateContractMutation(graphene.Mutation):
 
 
 class UpdateContractMutation(graphene.Mutation):
+    """
+    Mutation for updating an existing contract in the GraphQL API.
+
+    This mutation handles the update of a existing contract by accepting an input 
+    of type `ContractInput` and `graphene.ID`. It returns the updated contract object, a success flag, 
+    and a message indicating the result of the operation.
+    """
     class Arguments:
         id = graphene.ID(required=True)
         input = ContractInput(required=True)
@@ -141,6 +188,13 @@ class UpdateContractMutation(graphene.Mutation):
 
 
 class DeleteContractMutation(graphene.Mutation):
+    """
+    Mutation for deleting an existing contract in the GraphQL API.
+
+    This mutation handles the delete of a existing contract by accepting an input 
+    of type `graphene.ID`. It returns the updated user object, a success flag, 
+    and a message indicating the result of the operation.
+    """
     class Arguments:
         id = graphene.ID(required=True)
 
@@ -160,14 +214,22 @@ class DeleteContractMutation(graphene.Mutation):
 
 
 class Mutation(graphene.ObjectType):
+    """
+    The Mutation class represents all the queries that can perform
+    server-side data changes.
+    """
+
+    # authentication mutations
     token_auth = graphql_jwt.ObtainJSONWebToken.Field()
     verify_token = graphql_jwt.Verify.Field()
     refresh_token = graphql_jwt.Refresh.Field()
 
+    # user mutations
     create_user = CreateUserMutation.Field()
     update_user = UpdateUserMutation.Field()
     delete_user = DeleteUserMutation.Field()
 
+    # contract mutations
     create_contract = CreateContractMutation.Field()
     update_contract = UpdateContractMutation.Field()
     delete_contract = DeleteContractMutation.Field()
